backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.4
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
python-multipart==0.0.6
reportlab==4.0.7
pandas==2.1.3
openpyxl==3.1.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
channels==4.0.0
daphne==4.0.0 